+++ src/ttable.c
@@ -7,7 +7,7 @@
 extern volatile TTable ttable;
 
 
-static const uint64_t TTABLE_INIT_CAPACITY = 65536ULL; // Power of 2 for modulo efficiency
+static const size_t TTABLE_INIT_CAPACITY = 65536; // Power of 2 for modulo efficiency
 
 
 /**
